% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addPackage}
\alias{addPackage}
\title{Add packages to a miniCRAN repository}
\usage{
addPackage(pkgs, path = "./", repos = "https://cran.rstudio.com/")
}
\arguments{
\item{pkgs}{List of packages to add to your repository.}

\item{path}{Location of your miniCRAN.}

\item{repos}{CRAN url repository to use.}
}
\value{
Add selected packages and all the dependecies to your miniCRAN repository.
}
\description{
Add packages to a miniCRAN repository
}
\examples{
addPackage(c("dplyr","ggplot2"))
}

