% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{initialize_repo}
\alias{initialize_repo}
\title{Initiliazie a miniCRAN repository}
\usage{
initialize_repo(pkgs = c(), path = "./",
  repos = "https://cran.rstudio.com/")
}
\arguments{
\item{pkgs}{List of packages to add to your repository.}

\item{path}{Location of your miniCRAN.}

\item{repos}{CRAN url repository to use.}
}
\value{
Crete a CRAN like folder structure under \code{path}.
}
\description{
Initiliazie a miniCRAN repository
}
\examples{
initialize_repo()
}

