% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_repo}
\alias{update_repo}
\title{Update a miniCRAN repository}
\usage{
update_repo(path = "./", repos = "https://cran.rstudio.com/")
}
\arguments{
\item{path}{Location of your miniCRAN.}

\item{repos}{CRAN url repository to use.}
}
\value{
Update all the packages of a miniCRAN repository with the most recent version available on CRAN repo.
}
\description{
Update a miniCRAN repository
}
\examples{
update_repo()
}

